Add <uses-library> "org.apache.http.legacy" to Android.bp file.
The manifest for StorageManager contains a <uses-library> tag for
"org.apache.http.legacy", so Soong must know about this dependency in
order to pass verify_uses_libraries check and generate correct class
loader context for dexpreopt. Soong is unable to automatically deduce
the dependency from "libs" in this case, as it is not in "libs".

Bug: 132357300

Test: m out/soong/.intermediates/packages/apps/StorageManager/StorageManager/android_common/manifest_check/AndroidManifest.xml
Test: m out/soong/.intermediates/packages/apps/StorageManager/StorageManager/android_common/dexpreopt/oat/x86/package.odex && \
      oatdump --oat-file=out/soong/.intermediates/packages/apps/StorageManager/StorageManager/android_common/dexpreopt/oat/x86/package.odex 2>/dev/null | grep '^classpath'
      classpath = PCL[]{PCL[/system/framework/org.apache.http.legacy.jar*807226390]}

Change-Id: I53f4551f65cce0b0c743f4a0a0fb5bbef8c13364